MovieSearchPage*: restructurization; added router
